Reverted changes to attacking_direction_from_frame
The proposed solution on the pull request would be the correct one is
all teams had players, but at the moment that is only being done for the
EPTS dataset, as described in the previous comment on the pull request.

Till we do that, and since this method is only used for reading the
tracab and metrica data, we need to check if home or away is in the key
of the players_coordinates.