Added players (Player) to teams (Team) for tracab dataset
It ws tricky to implement because we don't know how many players there
are per team, nor what their jersey_no is till we start reading the
file. Thus, for each line I had to check if the player with that
jersey_no exist for each team, and if not create it an add it to the
team. I also created a method for Teams to be able to get the player by
jersey_no.

Not sure if there is a better way to implement this. Let me know!