Added serializer for Metrica event data in json format
In this commit we added the serializer and corresponding tests for
reading Metrica Sports event data in json format. The types and subtypes
provided by Metrica were used to extract the common events user for Opta
and Statsbomb event data [PASS, SHOT, CARRY, TAKE ON]. All the rest are
Generic Events.