Added improved standard events to Metrica serializer
As part of the larger issue of adding more standard event to be able to
better match the data format and event types between provier, i have
aded in this commit the following standard events:

	- GOAL KICK
	- THROW IN
	- FREE KICK
	- CORNER KICK
	- KICK OFF
	- PENALTY
	- RECOVERY
	- FOUL COMMITTED
	- BALL OUT

In the case of Metrica data, BALL OUT is a synthetic event that's added
to the dataset. All the other ones replace the orginal event, for
example, before a pass froma free kick was of event_type PASS, now it
woul dbe of event_type FREE KICK.

I just realized shots from penaly or free kicks are now correctly
handled, will fixed that in a next commit.

The same changes will also implemented in the other providers.