Qualifiers approach to events to code for things like set pieces
As discussed on #43, to be able to code for things like whether an event
is a set pieces, we moved from encoding that on the name of the event,
to adding qualifiers to the event. For now the qualifiers are only
SetPieceType, but could accomodate for other things like Foot, or
BodyPart for shots, etc.