Improve sequence definition based on latets added generic events
In this commit, as proposed in #72 I implemented and improvement on the
definition of sequences so that they are opened by a pass, carry or
recovery, and closed by a ball out, fould committed or shot on target.

This results also in sequences in which the team is None, whixh I think
is interesting so you can look at events or tracking data when no team
is in possession.