Improved state builder implementation: reduce_after and reduce_before
In some cases a state needs to be updated before the event (like in
recoveries, or passes) and in others after the event (like in score, or
ball out).

I have changed the single reduce instance which was taking place after
the event, by `redice_before` and `reduce_after` methods which allow
state builders to handle this two situations.