Fix sequences reset at set pieces
It is difficult to contemplate all the possible ways in which different
providers code for a ball out. This could lead to sequences not being
ended / started properly, leading to a se piece in the middle of a
sequence. In this commit we changed that so that we always start a new
sequence whenever a set piece event occurs.