DRAFT for default normalization of coordinates
In this commit I submit the a draft version of the default normalization
of coordiantes for the serializer of Tracab data. I submit this draft so
that we can discuss / modify / agree on an approach before implementing
it on the other serializers.

I have added a VerticalOrietnation model to keep track of the vertical
orientation of the dataset, as it can be reversed when going to original
to default and vice versa. This attribute is at the Metadata level, but
not sure if it should go on the Dimension or PitchDimensions model.

I added length and widht as attributes to the PitchDimensions model,
rather than them being a property, so that they can remain unchanged
when changing the coordinate system of the data.

In the serializer itself, I added the transformation on each frame
depending on the option given to it (normalize by default).

One question I have is whether we should keep a list of models /
coordiantes systems per provider, so that you can transform to a certain
coordiante system, or you can transform back after it being normalized,
for example `transform_to_coordiante_system(CoordianteSystem.TRACAB)`.
And then make Kloppy ber one of them and transform to that one by
default.