WIP creating coordinate system models and adapting transformers
We changed from setting the normalization inside each sealizer
independently to having models for each coordinate system type and then
modifiying the transformers for frames and datasets to be able to
transform from one coordiante system to another.

In this commit I implemented a coordiante system for kloppy and one for
Tracab, and modified the frame transformer.

Next step is to implement it for tracab as an example serializer.