Implemented feedback on PR.
We discussed on the PR to add a builder for the coordinates systems that
takes a str/name for the provider and any arguments needed to
instantiate the class.

You suggested leaving it as a str, so that users don't have to think
about the Provider Enum. While I agree that is more user friendly, I
believe that it could actually lead to errors as well. For example
giving as input `Tracab` instead of `tracab` and so on. The way I see it
the Provider Enum exist, and anyone uysing the package will benefitiate
from knowing about it. Also, the code is more clear.

Thus, I left here an implementation that takes the Provider as input. My
suggestion would be that if we think this could be problematic, we make
it clear and explain it in the documentation.

That said, in the end it's your call. If you want to do the `str`
implemetation is totally fine with me!