Addes standarization of coordinates to all serializers
In this commit we applied the standarization of coordinates to all
providers.

This PR is still work in progress as there are still tests to be
developed and features to be implemented. Like how to handle default
dimensions, don't transform when not necessary, make sure the
transformation of datasets works as well, etc.