Added tests and support for helpers and dataset transformers
In this commit we added test to check that all providers are in the
correct dimensinons when parsed to keep their original dimensions as
well as when they are normalized to the kloppy coordinate system.

Also I modified the transform helpers as well as the dataset transform
methods so that they are compatible with taking a coordiante system as
a target transformation.

An important issue here is that there can't be a transformation that
does at the same time a pitch dimensions and a coordinate system
transformation. Not sure this is handled the right way.

I would explore whether we actually need to have both, or else leave the
pitch dimensions inside the coordinate system (with a property in the
metadata) so that all the information about dimensions is contained
inside the coordinate system vs twice, one in the metadata attribute and
one in the coordiante system.