Changed MetricaTrackingSerializer to MetricaCsvSerializer
The MetricaTrackingSerilizer was the first serializer for Metrica Sports
tracking data included in kloppy. This serializers is for the CSV format
that Metrica Sports used to provide it's data in. It's also the format
of 2 of the 3 games with tracking that Metrica Sports made public.

Metrica Sports for quite some time now has a new format for their
tracking data. They now provide their tracking data in the EPTS FIFA
standard format. That data can be loaded with the EPTS selializer in
kloppy that will be now renamed to specify that that is the serializer
for Metrica Sports data.

Thus, to avoid confussion, I propose in this PR that we renamed the
original serializer and file to make it clear that those methods are for
the data in CSV format.

It's still pending any backwards compatibility or error we might need to
throw if users want to use the MetricaTrackingSeiliazer as it was
before.