Renamed everything related to EPTS serializer to MetricaEPTS
While a lot of the implementation of the EPTSSerializer would apply to
the general strcuture of the EPTS format, there are certain elements
like periods information or type of data that are specific to Metrica
Sports data.

There are no other providers at the moment which data is in EPTS
format. Thus, rather than create a base EPTSSerilizer and then do MetricaEPTS
serializer based on that one, for now we will rename the EPTSSerializer
to make clear that is a Metrica one.

We will refactor the code in the future if there is a need for another
EPTS serializer.