Changed from players_coordinates to players_data
Before for each player on each frame we were only serialing the x and y
coordinates of the tracking data. Most providers on top of that provide
the speed of each player on each frame, and some also provide the
distance covered.

In this first commit I changed the attribute in Frames from
players_coordiantes to players_data. players_data has the PlayerData
information for each player on that frame.

PlayerData is a new model that has coordinates, distance and speed
attributes.

This changes are in this commit only implemented for the Metrica Sports
EPTS serializer.

I also modified the helper to_pandas so that this extra data is included
on the resulting dataframe as well.