Added deserialization of non PlayerData to Metrica EPTS serializer
Depending on the data type, on Metrica Sports data there could be fields
that do not belong to PlayerData types (position, distance, speed). This
extra data types need to be deserialized, added to the Frame model on
other_data and then also included in the dataframe when using to_pandas
on the dataset.

The other_data is a dict where any kind of data can be included and it's
added to the Frame model. While for use now only on the Metrica EPTS
data it leaves a base for serialiazers of different providers to add
data other than the player and ball data.

Next step is to modify and changes tests and serializers for all other
providers.