Moved other players data to PlayerData model
In the first implementation of other_data we added it to the Frame
model. However, in the use cased needed for the Metrica EPTS format all
other data was associated with a specific player, thus it made more
sense to add it to the PlayerData model. In this PR we have moved it
there.

Since it could also be that there are other_data relevant at the frame
level, like for example the ball alive/dead of Tracab, we also left the
other_data at the Frame level as well.