Fix shot end coordinates for Opta deserializer
The end coordinates of blocked and saved shots were based on the projection of the shot on the goal mouth instead of the shot's actual end location.

This commit uses the x and y coordinates of the location were the shots was blocked (qualifiers 146 and 147) and inversely projects the goalmouth z  coordinate (qualifier 103) on the location were the shot was blocked.

Fixes #244