fix: Uniform implementation of timestamps
The representation of timestamps was inconsistent between different data providers.

This commit implements a uniform implementation between data providers. The timestamps of events and frames are now relative to the kick-off event in each period. This implies that timestamp are reset to zero for each period and that tracking frames before the kick-off event get a negative timestamp. The standard implementation for the start_timestamp and end_timestamp of periods is the absolute UTC datetime of respectively the kick-off and final whistle of the corresponding period. If the absolute datetime is not available, we use the offset between the start of the period's data feed and the kick-off for the start_timestamp, and the offset between the start of the period's data feed and the final whistle for the end_timestamp.

BREAKING CHANGE: The Period.start_timestamp and Period.end_timestamp fields now have type datetime or timedelta instead of float

BREAKING CHANGE: The DataRecord.timestamp field now has type timedelta instead of float

BREAKING CHANGE: The Period.contains method was removed

BREAKING CHANGE: The value of the Period.start_timestamp, Period.end_timestamp and DataRecord.timestamp fields changes for most dataproviders